Added a RivenInteractable actor.
This actor is initialized with a single mesh which it will animate between
a min a max value over a specified duration. The interactable is also given
a unique enumeration value corresponding to an in-game moveable object.
It will use this enumeration value when saving and restoring the actor
state in the game.

At present the game remembers the state, but does not persist it to disk.